Improved face culling + negative chunks
I made it so that checking blocks past a chunks bounds now checks the block in the next chunk over, meaning chunks now are almost one continuous mesh (visually). There used to be giant quads in between chunks, but that not longer exists. I also added negative chunks (chunks that exist on the negative x or z axis), and removed the check for if the global block position is less than zero. Also fixed some bugs.