Last commit (R.I.P.)
I have decided to hang up this project. It was fun, and I learned quite a bit. However, no matter what I do with optimizations and simplification, C# throws a tantrum and does 6+ms just to mesh a 16x250x16 chunk. That is awful times, even the redditors have .9ms meshing times. Imperceivable. So maybe one day I will come back. But for now, this is the last commit for this voxel project. I guess minecraft isn't so easy to make...